Move getBlocksFromTimestamps to separate util function
