+++ src/utils/index.js
@@ -15,6 +15,7 @@ export { default as dimensionsPropType } from './dimensionsPropType';
 export { default as directionPropType } from './directionPropType';
 export { default as ethereumUtils } from './ethereumUtils';
 export { default as gasUtils } from './gas';
+export { default as getBlocksFromTimestamps } from './getBlocksFromTimestamps';
 export { default as getDominantColorFromImage } from './getDominantColorFromImage';
 export { default as getTokenMetadata } from './getTokenMetadata';
 export { default as getUrlForTrustIconFallback } from './getUrlForTrustIconFallback';
